Day saving, photo moments, and app icon fixes
Icons: the installed app was showing Next's stock favicon.ico. Replaced
with icon.png and apple-icon.png generated from the logo, so the browser
tab, phone home screen and desktop install all carry the real mark.

Days now close. Today stays open; earlier days sit saved with a tick on the
calendar and read-only controls until reopened, then Save puts them away
again. Previous walks back indefinitely rather than stopping at the start
of the block.

Moments: up to nine photos a day as a 3-across grid, each opening full size
with a caption. Images are downscaled to 900px JPEG before storage — the
browser is holding these until accounts land, and a phone photo would
exhaust the quota on its own.

Undo drops sliders (trivially redraggable) and is now an always-present
arrow while editing, disabled when there's nothing to step back.

Daily Reflection is an always-open italic box in a branded card, matching
the habit comments rather than hiding behind an Add button.

Day view loses the "Check-in" subtitle, abbreviates the weekday, and the
calendar spells weekdays as Mon/Tue/Wed. Header logo grows to 56px with the
banner height held by trimming its padding.

+++ src/lib/habits.ts
@@ -83,11 +83,11 @@ function ordinal(day: number) {
   return ["th", "st", "nd", "rd"][day % 10] ?? "th";
 }
 
-// "Thursday 13th of August" — the day view's title, spelled out in full
-// rather than as a relative label like "Yesterday".
+// "Thu 13th of August" — the day view's title. Weekday abbreviated so the
+// line stays on one row between the Back and Previous/Next controls.
 export function fullDate(dayIndex: number) {
   const date = dateForDay(dayIndex);
-  const weekday = date.toLocaleDateString(undefined, { weekday: "long" });
+  const weekday = date.toLocaleDateString(undefined, { weekday: "short" });
   const month = date.toLocaleDateString(undefined, { month: "long" });
   const day = date.getDate();
   return `${weekday} ${day}${ordinal(day)} of ${month}`;
